refactor: add subcommand for each oracle type
