+++ cmd/oracle/ai/ai.go
@@ -1,0 +1,1 @@
+package aioracle
