+++ cmd/oracle/oracle.go
@@ -1,13 +1,8 @@
 package oracle
 
 import (
+	"github.com/dymensionxyz/roller/cmd/oracle/priceoracle"
 	"github.com/spf13/cobra"
-
-	"github.com/dymensionxyz/roller/cmd/services"
-	loadservices "github.com/dymensionxyz/roller/cmd/services/load"
-	restartservices "github.com/dymensionxyz/roller/cmd/services/restart"
-	startservices "github.com/dymensionxyz/roller/cmd/services/start"
-	stopservices "github.com/dymensionxyz/roller/cmd/services/stop"
 )
 
 func Cmd() *cobra.Command {
@@ -16,19 +11,7 @@ func Cmd() *cobra.Command {
 		Short: "Commands related to RollApp's component observability",
 	}
 
-	cmd.AddCommand(DeployCmd())
-	cmd.AddCommand(StartCmd())
-
-	sl := []string{"oracle"}
-	cmd.AddCommand(
-		services.Cmd(
-			loadservices.Cmd(sl, "oracle"),
-			startservices.OracleCmd(),
-			restartservices.Cmd(sl),
-			stopservices.Cmd(sl),
-			// logservices.RollappCmd(),
-		),
-	)
+	cmd.AddCommand(priceoracle.Cmd())
 
 	return cmd
 }
