+++ cmd/oracle/priceoracle/configs/wasm-config.yaml
