+++ cmd/oracle/priceoracle/deploy.go
@@ -1,4 +1,4 @@
-package oracle
+package priceoracle
 
 import (
 	"context"
@@ -14,6 +14,7 @@ import (
 
 	initconfig "github.com/dymensionxyz/roller/cmd/config/init"
 	"github.com/dymensionxyz/roller/cmd/consts"
+	oracleutils "github.com/dymensionxyz/roller/cmd/oracle/utils"
 	"github.com/dymensionxyz/roller/utils/config/yamlconfig"
 	"github.com/dymensionxyz/roller/utils/dependencies"
 	"github.com/dymensionxyz/roller/utils/filesystem"
@@ -21,13 +22,13 @@ import (
 	"github.com/dymensionxyz/roller/utils/roller"
 )
 
-//go:embed setup/configs/*
+//go:embed configs/*
 var configFiles embed.FS
 
 func DeployCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "deploy",
-		Short: "Deploys an oracle to the RollApp",
+		Short: "Deploys a price oracle to the RollApp",
 		Run: func(cmd *cobra.Command, args []string) {
 			if err := initconfig.AddFlags(cmd); err != nil {
 				pterm.Error.Printf("failed to add flags: %v\n", err)
@@ -54,11 +55,11 @@ func DeployCmd() *cobra.Command {
 			}
 
 			// Create the appropriate deployer based on RollApp type
-			var deployer ContractDeployer
+			var deployer oracleutils.ContractDeployer
 			var contractUrl string
 			switch rollerData.RollappVMType {
 			case consts.EVM_ROLLAPP:
-				deployer, err = NewEVMDeployer(rollerData)
+				deployer, err = oracleutils.NewEVMDeployer(rollerData)
 				if err != nil {
 					pterm.Error.Printf("failed to create evm deployer: %v\n", err)
 					return
@@ -71,7 +72,7 @@ func DeployCmd() *cobra.Command {
 					return
 				}
 			case consts.WASM_ROLLAPP:
-				deployer, err = NewWasmDeployer(rollerData)
+				deployer, err = oracleutils.NewWasmDeployer(rollerData)
 				if err != nil {
 					pterm.Error.Printf("failed to create wasm deployer: %v\n", err)
 					return
@@ -169,7 +170,7 @@ func DeployCmd() *cobra.Command {
 
 			switch rollerData.RollappVMType {
 			case consts.EVM_ROLLAPP:
-				networkID, err := extractNetworkID(rollerData.RollappID)
+				networkID, err := oracleutils.ExtractNetworkID(rollerData.RollappID)
 				if err != nil {
 					pterm.Error.Printf("failed to extract network ID: %v\n", err)
 					return
