+++ cmd/oracle/rng/rng.go
@@ -1,0 +1,1 @@
+package rngoracle
