+++ cmd/oracle/utils/deployer.go
@@ -1,4 +1,4 @@
-package oracle
+package oracleutils
 
 import (
 	"context"
