+++ cmd/oracle/utils/wasm_deployer.go
@@ -1,4 +1,4 @@
-package oracle
+package oracleutils
 
 import (
 	"context"
